refactor(session): colocate goals and document ownership rules
closes #2194
