+++ packages/coding-agent/src/modes/agent-connection/types.ts
@@ -14,7 +14,6 @@ import type {
 } from "../../core/cron-jobs.js";
 import type { ReplayBuiltInToolName } from "../../core/extensions/index.js";
 import type { InputSource } from "../../core/extensions/types.js";
-import type { GoalState } from "../../core/goals.js";
 import type { KernelSentAgentMessage } from "../../core/kernel/index.js";
 import type { AcpMcpServerConfig } from "../../core/mcp/acp-mcp-types.js";
 import type { RefinementResult } from "../../core/refinement/index.js";
@@ -28,6 +27,7 @@ import type {
 import type { DeleteSessionFileResult } from "../../core/session-file-actions.js";
 import type { SessionStats } from "../../core/session-stats.js";
 import type { SessionUsageSummary } from "../../core/usage.js";
+import type { GoalState } from "../../session/goals/contracts.js";
 import type { SessionSummary } from "../daemon/daemon-session-list.js";
 
 /**
