+++ packages/coding-agent/src/session/goals/controller.ts
@@ -1,6 +1,6 @@
 import { randomUUID } from "node:crypto";
 import type { AssistantMessage } from "@earendil-works/pi-ai";
-import { emptyGoalState, type GoalState, goalTokenDeltaForUsage, normalizeGoalState } from "../core/goals.js";
+import { emptyGoalState, type GoalState, goalTokenDeltaForUsage, normalizeGoalState } from "./contracts.js";
 import type { GoalPersistence } from "./persistence.js";
 
 export interface GoalCheckpoint {
