+++ packages/coding-agent/src/session/turns/command-execution.ts
@@ -2,7 +2,6 @@ import type { Agent } from "@earendil-works/pi-agent-core";
 import type { ImageContent } from "@earendil-works/pi-ai";
 import type { AgentSessionEvent } from "../../core/agent-session.js";
 import type { CompactionResult } from "../../core/compaction/index.js";
-import type { GoalState } from "../../core/goals.js";
 import {
 	type CustomMessage,
 	createSessionSlashCommandMessage,
@@ -18,6 +17,7 @@ import {
 import type { SessionManager } from "../../core/session-manager.js";
 import { parseRefineCommandOptions, type SessionSlashCommand } from "../../core/slash-commands.js";
 import { CompactionSkippedError } from "../compaction/compaction-execution.js";
+import type { GoalState } from "../goals/contracts.js";
 import type { SessionCommitFence, SessionCommitLease } from "../input/commit-fence.js";
 import type { QueuedSessionAction } from "../prepared-actions.js";
 import type { SessionRefinement } from "../refinement/refinement.js";
