+++ packages/coding-agent/src/session/turns/turn-policy.ts
@@ -7,8 +7,8 @@ import type { Api, AssistantMessage, Model } from "@earendil-works/pi-ai";
 import { shouldCompact } from "../../core/compaction/index.js";
 import { getLatestCompactionEntry, type SessionManager } from "../../core/session-manager.js";
 import type { SettingsManager } from "../../core/settings-manager.js";
-import type { GoalController } from "../../goals/controller.js";
 import type { SessionCompaction } from "../compaction/compaction.js";
+import type { GoalController } from "../goals/controller.js";
 import type { SessionRefinement } from "../refinement/refinement.js";
 import type { SessionAutonomousContinuation } from "./autonomous-continuation.js";
