+++ packages/coding-agent/test/goal-continuation-quiescence.test.ts
@@ -1,12 +1,12 @@
 import type { GetContinuationMessagesContext } from "@earendil-works/pi-agent-core";
 import { fauxAssistantMessage } from "@earendil-works/pi-ai";
 import { describe, expect, it, vi } from "vitest";
-import { emptyGoalState } from "../src/core/goals.js";
 import { ActionStore } from "../src/core/session-action-store.js";
-import { GoalController } from "../src/goals/controller.js";
+import { SessionGoalContinuation, type SessionGoalContinuationHost } from "../src/session/goals/continuation.js";
+import { emptyGoalState } from "../src/session/goals/contracts.js";
+import { GoalController } from "../src/session/goals/controller.js";
 import { SessionInputScheduler } from "../src/session/input/input-scheduler.js";
 import type { QueuedSessionAction } from "../src/session/prepared-actions.js";
-import { SessionGoalContinuation, type SessionGoalContinuationHost } from "../src/session/turns/goal-continuation.js";
 
 function harness(overrides: { awaitsChildWork?: boolean } & Partial<SessionGoalContinuationHost> = {}) {
 	const goals = new GoalController({ load: emptyGoalState, save: () => {} }, () => {});
