+++ packages/coding-agent/test/suite/regressions/4482-heartbeat-injected-prompt.test.ts
@@ -4,7 +4,6 @@ import { Container, type MarkdownTheme } from "@earendil-works/pi-tui";
 import { Type } from "typebox";
 import { afterEach, beforeAll, describe, expect, it, vi } from "vitest";
 import { type AgentCronJob, shouldDeferHeartbeatCronJob } from "../../../src/core/cron-jobs.js";
-import { createGoalContextMessage, type GoalState } from "../../../src/core/goals.js";
 import {
 	type CustomMessage,
 	createHeartbeatPromptMessage,
@@ -18,6 +17,7 @@ import {
 import { InteractiveMode } from "../../../src/modes/interactive/interactive-mode.js";
 import { getMarkdownTheme, initTheme } from "../../../src/modes/interactive/theme/theme.js";
 import type { SessionCompaction } from "../../../src/session/compaction/compaction.js";
+import { createGoalContextMessage, type GoalState } from "../../../src/session/goals/contracts.js";
 import { conversationMessages, createHarness, getMessageText, getUserTexts, type Harness } from "../harness.js";
 
 type AddMessageToChatHost = {
